SC-8029: refactoring validations
